pref: divide gnc.
